download changed, now you can see default picture if not connected

+++ system/application/helpers/actions_helper.php
@@ -56,7 +56,7 @@ function get_action_strings($actions, $tuple) {
 				$str .= '<li><a href="/bills/issue/'.$tuple['account_id'].'">Issue Bill</a></li>';
 				break;
 			case 'delete-bill':
-				$str .= '<li><a href="/bills/delete/'.$tuple['bill_id'].'">Delete Bill</a></li>';
+				$str .= '<li><a href="/bills/delete/'.$tuple['bill_id'].'" class="confirm">Delete Bill</a></li>';
 				break;
 			case 'pay-bill':
 				$str .= '<li><a href="/bills/pay/'.$tuple['bill_id'].'">Pay Bill</a></li>';
