+++ system/application/views/mainpane/lists/all_patients.php
@@ -28,12 +28,14 @@ if ($this->auth->get_type() == 'patient') {
 			<ul>
 				<li><a href="/profile/user/'.$table['tuples'][$i]['account_id'].'" class="ajax">See Profile</a></li>
 				<li><a href="/messages/compose/'.$table['tuples'][$i]['account_id'].'">Send Email</a></li>
-				<li><a href="/connections/destroy/'.$table['tuples'][$i]['account_id'].'">Delete Connection</a></li>
+				<li><a href="/connections/destroy/'.$table['tuples'][$i]['account_id'].'" class="confirm">Delete Connection</a></li>
 			</ul>';
 		} else {
-			$table['tuples'][$i]['*actions'] = '<a href="/connections/request/'.
-			$table['tuples'][$i]['account_id'].
-			'" class="confirm ajax">Request connection</a>';
+			$table['tuples'][$i]['*actions'] = '
+			<ul>
+				<li><a href="/profile/user/'.$table['tuples'][$i]['account_id'].'">See Profile</a></li>
+				<li><a href="/connections/request/'.$table['tuples'][$i]['account_id'].'" class="confirm">Request connection</a></li>
+			</ul>';
 		}
 	}
 } else { // You are an HCP (you cannot request connection to a patient)
@@ -41,11 +43,15 @@ if ($this->auth->get_type() == 'patient') {
 		if ($table['tuples'][$i]['connected']) {
 			$table['tuples'][$i]['*actions'] = '
 			<ul>
+				<li><a href="/profile/user/'.$table['tuples'][$i]['account_id'].'">See Profile</a></li>
 				<li><a href="/messages/compose/'.$table['tuples'][$i]['account_id'].'">Send Email</a></li>
-				<li><a href="/connections/destroy/'.$table['tuples'][$i]['account_id'].'">Delete Connection</a></li>
+				<li><a href="/connections/destroy/'.$table['tuples'][$i]['account_id'].'" class="confirm">Delete Connection</a></li>
 			</ul>';
 		} else {
-			$table['tuples'][$i]['*actions'] = '';
+			$table['tuples'][$i]['*actions'] = '
+			<ul>
+				<li><a href="/profile/user/'.$table['tuples'][$i]['account_id'].'">See Profile</a></li>
+			</ul>';
 		}
 	}
 }
